backend/requirements.txt
fastapi==0.115.6
httpx==0.28.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
uvicorn==0.34.0
